more preparation for new middleware
