+++ test/kernel/kernel.test.ts
@@ -18,59 +18,6 @@ describe("Kernel", () => {
         sandbox.restore();
     });
 
-    it("Should be able to use middleware as configuration", () => {
-
-        function logger(next: (context: IContext) => any) {
-            return (context: IContext) => {
-                return next(context);
-            };
-        };
-
-        let kernel = new Kernel();
-        kernel.applyMiddleware(logger);
-        let _kernel: any = kernel;
-        expect(_kernel._middleware).not.to.eql(null);
-
-    });
-
-    it("Should invoke middleware", () => {
-
-        interface INinja {}
-
-        @injectable()
-        class Ninja implements INinja {}
-
-        let log: string[] = [];
-
-        function middleware1(next: (context: IContext) => any) {
-            return (context: IContext) => {
-                let serviceIdentifier = context.kernel.getServiceIdentifierAsString(context.plan.rootRequest.serviceIdentifier);
-                log.push(`Middleware1: ${serviceIdentifier}`);
-                return next(context);
-            };
-        };
-
-        function middleware2(next: (context: IContext) => any) {
-            return (context: IContext) => {
-                let serviceIdentifier = context.kernel.getServiceIdentifierAsString(context.plan.rootRequest.serviceIdentifier);
-                log.push(`Middleware2: ${serviceIdentifier}`);
-                return next(context);
-            };
-        };
-
-        let kernel = new Kernel();
-        kernel.applyMiddleware(middleware1, middleware2);
-        kernel.bind<INinja>("INinja").to(Ninja);
-
-        let ninja = kernel.get<INinja>("INinja");
-
-        expect(ninja instanceof Ninja).eql(true);
-        expect(log.length).eql(2);
-        expect(log[0]).eql(`Middleware1: INinja`);
-        expect(log[1]).eql(`Middleware2: INinja`);
-
-    });
-
   it("Should be able to use modules as configuration", () => {
 
       interface INinja {}
